+++ src/Compute/Compute.Test/ScenarioTests/StrategiesVmssTests.cs
@@ -137,5 +137,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-SimpleNewVmssSkipExtOverprovision");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestSimpleNewVmssDeprecatedImageWarning()
+        {
+            TestRunner.RunTestScript("Test-SimpleNewVmssDeprecatedImageWarning");
+        }
     }
 }